workers/requirements.txt
pika
psycopg2-binary
transformers
torch
pysentimiento
requests
sentence-transformers
numpy

